+++ components/BottomNavigation/src/private/MDCBottomNavigationItemBadge.m
@@ -16,19 +16,6 @@
 
 #import "MDCBottomNavigationItemBadge.h"
 
-static NSString *const kMDCBottomNavigationItemBadgeCircleWidthKey =
-    @"kMDCBottomNavigationItemBadgeCircleWidthKey";
-static NSString *const kMDCBottomNavigationItemBadgeCircleHeightKey =
-    @"kMDCBottomNavigationItemBadgeCircleHeightKey";
-static NSString *const kMDCBottomNavigationItemBadgeXPaddingKey =
-    @"kMDCBottomNavigationItemBadgeXPaddingKey";
-static NSString *const kMDCBottomNavigationItemBadgeYPaddingKey =
-    @"kMDCBottomNavigationItemBadgeYPaddingKey";
-static NSString *const kMDCBottomNavigationItemBadgeValueKey =
-    @"kMDCBottomNavigationItemBadgeValueKey";
-static NSString *const kMDCBottomNavigationItemBadgeColorKey =
-    @"kMDCBottomNavigationItemBadgeColorKey";
-
 static const CGFloat kMDCBottomNavigationItemBadgeFontSize = 10.f;
 static const CGFloat kMDCBottomNavigationItemBadgeXPadding = 8.f;
 static const CGFloat kMDCBottomNavigationItemBadgeYPadding = 2.f;
@@ -46,46 +33,9 @@ static const CGFloat kMDCBottomNavigationItemBadgeYPadding = 2.f;
 - (instancetype)initWithCoder:(NSCoder *)aDecoder {
   self = [super initWithCoder:aDecoder];
   if (self) {
-    if ([aDecoder containsValueForKey:kMDCBottomNavigationItemBadgeCircleWidthKey]) {
-      _badgeCircleWidth =
-          (CGFloat)[aDecoder decodeDoubleForKey:kMDCBottomNavigationItemBadgeCircleWidthKey];
-    }
-    if ([aDecoder containsValueForKey:kMDCBottomNavigationItemBadgeCircleHeightKey]) {
-      _badgeCircleHeight =
-          (CGFloat)[aDecoder decodeDoubleForKey:kMDCBottomNavigationItemBadgeCircleHeightKey];
-    }
-    if ([aDecoder containsValueForKey:kMDCBottomNavigationItemBadgeXPaddingKey]) {
-      _xPadding = (CGFloat)[aDecoder decodeDoubleForKey:kMDCBottomNavigationItemBadgeXPaddingKey];
-    }
-    if ([aDecoder containsValueForKey:kMDCBottomNavigationItemBadgeYPaddingKey]) {
-      _yPadding = (CGFloat)[aDecoder decodeDoubleForKey:kMDCBottomNavigationItemBadgeYPaddingKey];
-    }
-    if ([aDecoder containsValueForKey:kMDCBottomNavigationItemBadgeValueKey]) {
-      _badgeValue = [aDecoder decodeObjectOfClass:[NSString class]
-                                           forKey:kMDCBottomNavigationItemBadgeValueKey];
-    }
-    if ([aDecoder containsValueForKey:kMDCBottomNavigationItemBadgeColorKey]) {
-      _badgeColor = [aDecoder decodeObjectOfClass:[UIColor class]
-                                           forKey:kMDCBottomNavigationItemBadgeColorKey];
-    }
-
     [self commonMDCBottomNavigationItemBadgeInit];
   }
   return self;
-}
-
-- (void)encodeWithCoder:(NSCoder *)aCoder {
-  [super encodeWithCoder:aCoder];
-  [aCoder encodeDouble:self.badgeCircleWidth forKey:kMDCBottomNavigationItemBadgeCircleWidthKey];
-  [aCoder encodeDouble:self.badgeCircleHeight forKey:kMDCBottomNavigationItemBadgeCircleHeightKey];
-  [aCoder encodeDouble:self.xPadding forKey:kMDCBottomNavigationItemBadgeXPaddingKey];
-  [aCoder encodeDouble:self.yPadding forKey:kMDCBottomNavigationItemBadgeYPaddingKey];
-  if (self.badgeValue != nil) {
-    [aCoder encodeObject:self.badgeValue forKey:kMDCBottomNavigationItemBadgeValueKey];
-  }
-  if (self.badgeColor != nil) {
-    [aCoder encodeObject:self.badgeColor forKey:kMDCBottomNavigationItemBadgeColorKey];
-  }
 }
 
 - (void)commonMDCBottomNavigationItemBadgeInit {