+++ components/FlexibleHeader/examples/FlexibleHeaderConfiguratorExample.m
@@ -19,7 +19,7 @@
 #import "MaterialFlexibleHeader.h"
 #import "supplemental/FlexibleHeaderConfiguratorSupplemental.h"
 
-@interface FlexibleHeaderConfiguratorExample () <MDCFlexibleHeaderViewLayoutDelegate>
+@interface FlexibleHeaderConfiguratorExample ()
 @property(nonatomic) BOOL overrideStatusBarHidden;
 @end
 
@@ -144,26 +144,6 @@
   if (scrollView == headerView.trackingScrollView) {
     [headerView trackingScrollViewWillEndDraggingWithVelocity:velocity
                                           targetContentOffset:targetContentOffset];
-  }
-}
-
-#pragma mark - MDCFlexibleHeaderViewLayoutDelegate
-
-- (void)flexibleHeaderViewController:
-            (nonnull MDCFlexibleHeaderViewController *)flexibleHeaderViewController
-    flexibleHeaderViewFrameDidChange:(nonnull MDCFlexibleHeaderView *)flexibleHeaderView {
-  CGFloat headerContentAlpha;
-  switch (flexibleHeaderView.scrollPhase) {
-    case MDCFlexibleHeaderScrollPhaseCollapsing:
-    case MDCFlexibleHeaderScrollPhaseOverExtending:
-      headerContentAlpha = 1;
-      break;
-    case MDCFlexibleHeaderScrollPhaseShifting:
-      headerContentAlpha = 1 - flexibleHeaderView.scrollPhasePercentage;
-      break;
-  }
-  for (UIView *subview in self.fhvc.headerView.subviews) {
-    subview.alpha = headerContentAlpha;
   }
 }