+++ components/FlexibleHeader/src/MDCFlexibleHeaderContainerViewController.m
@@ -25,6 +25,7 @@
   self = [super initWithNibName:nil bundle:nil];
   if (self) {
     _headerViewController = [[MDCFlexibleHeaderViewController alloc] init];
+    _headerViewController.inferTopSafeAreaInsetFromViewController = YES;
     [self addChildViewController:_headerViewController];
 
     self.contentViewController = contentViewController;