+++ components/FlexibleHeader/src/MDCFlexibleHeaderView.m
@@ -19,10 +19,14 @@
 #import "MaterialApplication.h"
 #import "MaterialUIMetrics.h"
 #import "private/MDCStatusBarShifter.h"
+#import "private/MDCFlexibleHeaderView+Private.h"
 
 #if TARGET_IPHONE_SIMULATOR
 float UIAnimationDragCoefficient(void);  // Private API for simulator animation speed
 #endif
+
+// The default status bar height for non-X devices.
+static const CGFloat kNonXStatusBarHeight = 20;
 
 // The default maximum height for the header. Does not include the status bar height.
 static const CGFloat kFlexibleHeaderDefaultHeight = 56;
@@ -74,6 +78,16 @@ static inline MDCFlexibleHeaderShiftBehavior ShiftBehaviorForCurrentAppContext(
 // to check what the intensity of a custom shadow should be depending on a scroll position. Valid
 // values range from 0 to 1. Where 0 is no shadow is visible and 1 is the shadow is fully visible.
 @property(nonatomic, readonly) CGFloat shadowIntensity;
+
+// The top safe area inset to consider when calculating the true minimum and maximum height of the
+// flexible header view.
+//
+// This property is typically assigned the value of view.safeAreaInsets.top (iOS 11 and up) or
+// viewController.topLayoutGuide.length (below iOS 11) from the header view controller's parent view
+// controller.
+//
+// This property is ignored if inferTopSafeAreaInsetFromViewController is NO.
+@property(nonatomic) CGFloat topSafeAreaInset;
 
 @end
 
@@ -130,6 +144,11 @@ static inline MDCFlexibleHeaderShiftBehavior ShiftBehaviorForCurrentAppContext(
   BOOL _interfaceOrientationIsChanging;
   BOOL _contentInsetsAreChanging;
 
+  // When the user tosses the scroll view we enter a deceleration phase. This is always eventually
+  // followed up by a call to trackingScrollViewDidEndDecelerating. See the
+  // trackingScrollViewDidEndDecelerating for more details on this behavior.
+  BOOL _didDecelerate;
+
   // _isChangingStatusBarVisibility documents whether we know that we're adjusting the status bar
   // visibility, while _wasStatusBarHidden allows us to detect whether someone else has adjusted
   // the status bar visibility. In either case, we need to counteract any content offsets
@@ -138,6 +157,13 @@ static inline MDCFlexibleHeaderShiftBehavior ShiftBehaviorForCurrentAppContext(
   BOOL _isChangingStatusBarVisibility;
   BOOL _wasStatusBarHiddenIsValid;
   BOOL _wasStatusBarHidden;
+
+  // We use this value to correctly handle top safe area insets on non-iPhone X devices.
+  CGFloat _lastNonZeroTopSafeAreaInset;
+
+  // UILayoutGuide was introduced in iOS 9, so in order to support iOS 8, we use a UIView as a
+  // layout guide. Once we drop iOS 8 support this can be changed to a UILayoutGuide instead.
+  UIView *_topSafeAreaGuide;
 
   MDCStatusBarShifter *_statusBarShifter;
 
@@ -168,6 +194,9 @@ static inline MDCFlexibleHeaderShiftBehavior ShiftBehaviorForCurrentAppContext(
 @synthesize inFrontOfInfiniteContent = _inFrontOfInfiniteContent;
 @synthesize sharedWithManyScrollViews = _sharedWithManyScrollViews;
 @synthesize visibleShadowOpacity = _visibleShadowOpacity;
+
+// Private
+@synthesize topSafeAreaInset = _topSafeAreaInset;
 
 - (void)dealloc {
 #if DEBUG
@@ -213,7 +242,7 @@ static inline MDCFlexibleHeaderShiftBehavior ShiftBehaviorForCurrentAppContext(
   _statusBarHintCanOverlapHeader = YES;
 
   _minMaxHeightIncludesSafeArea = YES;
-  _minimumHeight = kFlexibleHeaderDefaultHeight + MDCDeviceTopSafeAreaInset();
+  _minimumHeight = kFlexibleHeaderDefaultHeight + [self fhv_topSafeAreaInset];
   _maximumHeight = _minimumHeight;
 
   _visibleShadowOpacity = kDefaultVisibleShadowOpacity;
@@ -231,6 +260,10 @@ static inline MDCFlexibleHeaderShiftBehavior ShiftBehaviorForCurrentAppContext(
   _customShadowLayer = [CALayer layer];
   _customShadowLayer.hidden = YES;
   [self.layer addSublayer:_customShadowLayer];
+
+  _topSafeAreaGuide = [[UIView alloc] init];
+  _topSafeAreaGuide.frame = CGRectMake(0, 0, 0, [self fhv_topSafeAreaInset]);
+  [self addSubview:_topSafeAreaGuide];
 
   _contentView = [[UIView alloc] initWithFrame:self.bounds];
   _contentView.autoresizingMask =
@@ -332,6 +365,31 @@ static inline MDCFlexibleHeaderShiftBehavior ShiftBehaviorForCurrentAppContext(
   [_statusBarShifter didMoveToWindow];
 }
 
+- (void)extractTopSafeAreaInset {
+  UIViewController *viewController = self.topSafeAreaSourceViewController;
+  if (![viewController isViewLoaded]) {
+    self.topSafeAreaInset = 0;
+  } else {
+#if defined(__IPHONE_11_0) && (__IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_11_0)
+    if (@available(iOS 11.0, *)) {
+      self.topSafeAreaInset = viewController.view.safeAreaInsets.top;
+    } else {
+      self.topSafeAreaInset = viewController.topLayoutGuide.length;
+    }
+#else
+    self.topSafeAreaInset = viewController.topLayoutGuide.length;
+#endif
+  }
+}
+
+- (void)setTopSafeAreaSourceViewController:(UIViewController *)topSafeAreaSourceViewController {
+  _topSafeAreaSourceViewController = topSafeAreaSourceViewController;
+
+  if (self.inferTopSafeAreaInsetFromViewController) {
+    [self extractTopSafeAreaInset];
+  }
+}
+
 - (UIView *)hitTest:(CGPoint)point withEvent:(UIEvent *)event {
   UIView *hitView = [super hitTest:point withEvent:event];
 
@@ -362,36 +420,120 @@ static inline MDCFlexibleHeaderShiftBehavior ShiftBehaviorForCurrentAppContext(
 - (void)safeAreaInsetsDidChange {
 #if defined(__IPHONE_11_0) && (__IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_11_0)
   if (@available(iOS 11.0, *)) {
-    // If the min or max height have been explicitly set, don't adjust anything if the values
-    // already include a Safe Area inset.
-    BOOL hasSetMinOrMaxHeight = _hasExplicitlySetMinHeight || _hasExplicitlySetMaxHeight;
-    if (!hasSetMinOrMaxHeight && _minMaxHeightIncludesSafeArea) {
-      // If we're using the defaults we need to update them to account for the new Safe Area inset.
-      _minimumHeight = kFlexibleHeaderDefaultHeight + MDCDeviceTopSafeAreaInset();
-      _maximumHeight = _minimumHeight;
-    }
+    [super safeAreaInsetsDidChange];
 
-    // Adjust the scroll view insets to account for the new Safe Area inset.
-    [self fhv_enforceInsetsForScrollView:_trackingScrollView];
+    if (self.inferTopSafeAreaInsetFromViewController) {
+      // Will eventually invoke fhv_topSafeAreaInsetDidChange if needed.
+      [self extractTopSafeAreaInset];
 
-    // Ignore any content offset delta that occured as a result of any safe area insets change.
-    _shiftAccumulatorLastContentOffset = [self fhv_boundedContentOffset];
-
-    // The changes might require us to re-calculate the frame, or update the entire layout.
-    if (!_trackingScrollView) {
-      CGRect bounds = self.bounds;
-      bounds.size.height = self.computedMinimumHeight;
-      self.bounds = bounds;
-      CGPoint position = self.center;
-      position.y = -MIN([self fhv_accumulatorMax], _shiftAccumulator);
-      position.y += self.bounds.size.height / 2;
-      self.center = position;
-      [self.delegate flexibleHeaderViewFrameDidChange:self];
     } else {
-      [self fhv_updateLayout];
+      [self fhv_topSafeAreaInsetDidChange];
     }
   }
 #endif
+}
+
+#pragma mark - Top Safe Area Inset
+
+- (void)fhv_topSafeAreaInsetDidChange {
+  // If the min or max height have been explicitly set, don't adjust anything if the values
+  // already include a Safe Area inset.
+  BOOL hasSetMinOrMaxHeight = _hasExplicitlySetMinHeight || _hasExplicitlySetMaxHeight;
+  if (!hasSetMinOrMaxHeight && _minMaxHeightIncludesSafeArea) {
+    // If we're using the defaults we need to update them to account for the new Safe Area inset.
+    _minimumHeight = kFlexibleHeaderDefaultHeight + [self fhv_topSafeAreaInset];
+    _maximumHeight = _minimumHeight;
+  }
+
+  _topSafeAreaGuide.frame = CGRectMake(0, 0, self.bounds.size.width, [self fhv_topSafeAreaInset]);
+
+  // Adjust the scroll view insets to account for the new Safe Area inset.
+  [self fhv_enforceInsetsForScrollView:_trackingScrollView];
+
+  // Ignore any content offset delta that occured as a result of any safe area insets change.
+  _shiftAccumulatorLastContentOffset = [self fhv_boundedContentOffset];
+
+  // The changes might require us to re-calculate the frame, or update the entire layout.
+  if (!_trackingScrollView) {
+    CGRect bounds = self.bounds;
+    bounds.size.height = self.computedMinimumHeight;
+    self.bounds = bounds;
+    CGPoint position = self.center;
+    position.y = -MIN([self fhv_accumulatorMax], _shiftAccumulator);
+    position.y += self.bounds.size.height / 2;
+    self.center = position;
+    [self.delegate flexibleHeaderViewFrameDidChange:self];
+  } else {
+    [self fhv_updateLayout];
+  }
+}
+
+// Generally-speaking, we consider the top safe area inset to equate to the length of any
+// "device-owned" pixels.
+//
+// On an iPhone X, the top safe area inset is a fixed hardware constant, varying only based on
+// device orientation.
+//
+// On non-X devices, however, the top safe area inset is a flexible software value reflecting the
+// status bar's current height. If the status bar is hidden, the top safe area inset is zero.
+//
+// This affects the flexible header mostly when the status bar shift behavior is enabled. The
+// flexible header is able to hide the status bar interactively, meaning our top safe area inset
+// can fluctuate between 0 and 20 on non-X devices as we show/hide the status bar. This can cause
+// the tracking scroll view's content inset to "jump" because we make the scroll view's top
+// content inset equal maximumHeight + topSafeAreaInset.
+//
+// Aside: If we only supported iOS 11 and up we'd likely be able to simplify a lot of this
+// behavior by solely relying on additionalSafeAreaInsets. Alas, that will likely be sometime
+// after 2020.
+//
+// So, to avoid the jumping behavior, we keep track of what the last non-zero top safe area inset
+// was. If it was 20 and we know we've programmatically hidden the status bar, we continue to
+// pretend that the top safe area inset is 20. Once the status bar is visible again we rely on the
+// actual topSafeAreaInset value.
+- (CGFloat)fhv_topSafeAreaInset {
+  if (self.inferTopSafeAreaInsetFromViewController) {
+    BOOL topSafeAreaInsetLikelyAffectedByStatusBarVisibility =
+        _lastNonZeroTopSafeAreaInset == kNonXStatusBarHeight;
+    if (topSafeAreaInsetLikelyAffectedByStatusBarVisibility
+        && [self fhv_canShiftOffscreen]
+        && _shiftBehavior == MDCFlexibleHeaderShiftBehaviorEnabledWithStatusBar
+        && _statusBarShifter.prefersStatusBarHidden) {
+      return _lastNonZeroTopSafeAreaInset;
+
+    } else {
+      return self.topSafeAreaInset;
+    }
+  } else {
+    return MDCDeviceTopSafeAreaInset();
+  }
+}
+
+- (void)setTopSafeAreaInset:(CGFloat)topSafeAreaInset {
+  _topSafeAreaInset = topSafeAreaInset;
+
+  if (_topSafeAreaInset > 0) {
+    _lastNonZeroTopSafeAreaInset = _topSafeAreaInset;
+  }
+
+  [self fhv_topSafeAreaInsetDidChange];
+}
+
+- (void)setInferTopSafeAreaInsetFromViewController:(BOOL)inferTopSafeAreaInsetFromViewController {
+  if (_inferTopSafeAreaInsetFromViewController == inferTopSafeAreaInsetFromViewController) {
+    return;
+  }
+  _inferTopSafeAreaInsetFromViewController = inferTopSafeAreaInsetFromViewController;
+
+  if (_inferTopSafeAreaInsetFromViewController) {
+    [self extractTopSafeAreaInset];
+  } else {
+    [self fhv_topSafeAreaInsetDidChange];
+  }
+}
+
+- (id)topSafeAreaGuide {
+  return _topSafeAreaGuide;
 }
 
 #pragma mark - Private (fhv_ prefix)
@@ -468,7 +610,7 @@ static inline MDCFlexibleHeaderShiftBehavior ShiftBehaviorForCurrentAppContext(
 //
 // Our desired top content inset is always at least:
 //
-//     _maximumHeight (with safe area insets removed) + MDCDeviceTopSafeAreaInset()
+//     _maximumHeight (with safe area insets removed) + [self fhv_topSafeAreaInset]
 //
 // This ensures that when our scroll view is scrolled to its top that our header is able to be fully
 // expanded.
@@ -495,7 +637,7 @@ static inline MDCFlexibleHeaderShiftBehavior ShiftBehaviorForCurrentAppContext(
   // we ensure that our desired top inset is always at least the header height.
   CGFloat minimumTopInset;
   if (_minMaxHeightIncludesSafeArea) {
-    minimumTopInset = _maximumHeight - MDCDeviceTopSafeAreaInset();
+    minimumTopInset = _maximumHeight - [self fhv_topSafeAreaInset];
   } else {
     minimumTopInset = _maximumHeight;
   }
@@ -812,7 +954,7 @@ static inline MDCFlexibleHeaderShiftBehavior ShiftBehaviorForCurrentAppContext(
   // here.
   BOOL hasSetMinOrMaxHeight = _hasExplicitlySetMinHeight || _hasExplicitlySetMaxHeight;
   if (!hasSetMinOrMaxHeight && _minMaxHeightIncludesSafeArea) {
-    _minimumHeight = kFlexibleHeaderDefaultHeight + MDCDeviceTopSafeAreaInset();
+    _minimumHeight = kFlexibleHeaderDefaultHeight + [self fhv_topSafeAreaInset];
     _maximumHeight = _minimumHeight;
   }
 
@@ -961,6 +1103,10 @@ static inline MDCFlexibleHeaderShiftBehavior ShiftBehaviorForCurrentAppContext(
 #if DEBUG
   _didAdjustTargetContentOffset = NO;
 #endif
+
+  if (self.trackingScrollView.isTracking) {
+    _didDecelerate = NO; // Invalidate the flag - we're not actually decelerating right now.
+  }
 
   // We generally expect the tracking scroll view to be a sibling to the flexible header, but there
   // are cases where this assumption is always incorrect.
@@ -1142,9 +1288,20 @@ static BOOL isRunningiOS10_3OrAbove() {
   if (!willDecelerate && [self fhv_isPartiallyShifted]) {
     [self fhv_startDisplayLink];
   }
+  _didDecelerate = willDecelerate;
 }
 
 - (void)trackingScrollViewDidEndDecelerating {
+  // This event can be invoked after two different forms of user interaction:
+  //
+  // 1. When the tracking scroll view was tossed and then came to rest.
+  // 2. When the tracking scroll view was tossed, then grabbed and released (without another toss).
+  //
+  // We only want to react to the first type of interaction. _didDecelerate will only be true in the
+  // first case.
+  if (!_didDecelerate) {
+    return;
+  }
   if ([self fhv_isPartiallyShifted]) {
     _wantsToBeHidden =
         (_shiftAccumulator >= (1 - kMinimumVisibleProportion) * [self fhv_accumulatorMax]);
@@ -1225,6 +1382,9 @@ static BOOL isRunningiOS10_3OrAbove() {
   NSAssert(_interfaceOrientationIsChanging, @"Call to %@::%@ not matched by a call to %@.",
            NSStringFromClass([self class]), NSStringFromSelector(_cmd),
            NSStringFromSelector(@selector(interfaceOrientationWillChange)));
+  if (self.inferTopSafeAreaInsetFromViewController) {
+    [self extractTopSafeAreaInset];
+  }
   [self fhv_updateLayout];
 }
 
@@ -1307,7 +1467,7 @@ static BOOL isRunningiOS10_3OrAbove() {
   if (_minMaxHeightIncludesSafeArea) {
     return _minimumHeight;
   } else {
-    return _minimumHeight + MDCDeviceTopSafeAreaInset();
+    return _minimumHeight + [self fhv_topSafeAreaInset];
   }
 }
 
@@ -1315,7 +1475,7 @@ static BOOL isRunningiOS10_3OrAbove() {
   if (_minMaxHeightIncludesSafeArea) {
     return _maximumHeight;
   } else {
-    return _maximumHeight + MDCDeviceTopSafeAreaInset();
+    return _maximumHeight + [self fhv_topSafeAreaInset];
   }
 }
 
@@ -1328,14 +1488,14 @@ static BOOL isRunningiOS10_3OrAbove() {
   // Update default values accordingly.
   if (!_hasExplicitlySetMinHeight) {
     if (_minMaxHeightIncludesSafeArea) {
-      _minimumHeight = kFlexibleHeaderDefaultHeight + MDCDeviceTopSafeAreaInset();
+      _minimumHeight = kFlexibleHeaderDefaultHeight + [self fhv_topSafeAreaInset];
     } else {
       _minimumHeight = kFlexibleHeaderDefaultHeight;
     }
   }
   if (!_hasExplicitlySetMaxHeight) {
     if (_minMaxHeightIncludesSafeArea) {
-      _maximumHeight = kFlexibleHeaderDefaultHeight + MDCDeviceTopSafeAreaInset();
+      _maximumHeight = kFlexibleHeaderDefaultHeight + [self fhv_topSafeAreaInset];
     } else {
       _maximumHeight = kFlexibleHeaderDefaultHeight;
     }
@@ -1411,6 +1571,14 @@ static BOOL isRunningiOS10_3OrAbove() {
     // that the header's height won't change.
     CGPoint offset = scrollView.contentOffset;
     offset.y = self.trackingScrollView.contentOffset.y;
+    scrollView.contentOffset = offset;
+
+  } else if (self.trackingScrollView.contentOffset.y > scrollView.contentOffset.y
+             && scrollView.contentOffset.y < 0) { // Destination is showing an expanded header.
+    // Our header is possibly smaller now than it will be when we move to the new content.
+    // Scroll the new content such that we collapse the header to the current height.
+    CGPoint offset = scrollView.contentOffset;
+    offset.y = MIN(self.trackingScrollView.contentOffset.y, 0);
     scrollView.contentOffset = offset;
   }
 }