+++ components/TextFields/src/MDCTextInputControllerBase.m
@@ -43,63 +43,6 @@ static const NSTimeInterval
 static const NSTimeInterval
     MDCTextInputControllerBaseDefaultFloatingPlaceholderUpAnimationDuration = 0.3f;
 
-static NSString *const MDCTextInputControllerBaseActiveColorKey =
-    @"MDCTextInputControllerBaseActiveColorKey";
-static NSString *const MDCTextInputControllerBaseBorderFillColorKey =
-    @"MDCTextInputControllerBaseBorderFillColorKey";
-static NSString *const MDCTextInputControllerBaseCharacterCounterKey =
-    @"MDCTextInputControllerBaseCharacterCounterKey";
-static NSString *const MDCTextInputControllerBaseCharacterCountViewModeKey =
-    @"MDCTextInputControllerBaseCharacterCountViewModeKey";
-static NSString *const MDCTextInputControllerBaseCharacterCountMaxKey =
-    @"MDCTextInputControllerBaseCharacterCountMaxKey";
-static NSString *const MDCTextInputControllerBaseRoundedCorners =
-    @"MDCTextInputControllerBaseRoundedCorners";
-static NSString *const MDCTextInputControllerBaseDisabledColorKey =
-    @"MDCTextInputControllerBaseDisabledColorKey";
-static NSString *const MDCTextInputControllerBaseErrorAccessibilityValueKey =
-    @"MDCTextInputControllerBaseErrorAccessibilityValueKey";
-static NSString *const MDCTextInputControllerBaseErrorColorKey =
-    @"MDCTextInputControllerBaseErrorColorKey";
-static NSString *const MDCTextInputControllerBaseErrorTextKey =
-    @"MDCTextInputControllerBaseErrorTextKey";
-static NSString *const MDCTextInputControllerBaseExpandsOnOverflowKey =
-    @"MDCTextInputControllerBaseExpandsOnOverflowKey";
-static NSString *const MDCTextInputControllerBaseFloatingEnabledKey =
-    @"MDCTextInputControllerBaseFloatingEnabledKey";
-static NSString *const MDCTextInputControllerBaseFloatingPlaceholderNormalColorKey =
-    @"MDCTextInputControllerBaseFloatingPlaceholderNormalColorKey";
-static NSString *const MDCTextInputControllerBaseFloatingPlaceholderScaleKey =
-    @"MDCTextInputControllerBaseFloatingPlaceholderScaleKey";
-static NSString *const MDCTextInputControllerBaseHelperTextKey =
-    @"MDCTextInputControllerBaseHelperTextKey";
-static NSString *const MDCTextInputControllerBaseInlinePlaceholderFontKey =
-    @"MDCTextInputControllerBaseInlinePlaceholderFontKey";
-static NSString *const MDCTextInputControllerBaseInlinePlaceholderColorKey =
-    @"MDCTextInputControllerBaseInlinePlaceholderColorKey";
-static NSString *const MDCTextInputControllerBaseLeadingUnderlineLabelFontKey =
-    @"MDCTextInputControllerBaseLeadingUnderlineLabelFontKey";
-static NSString *const MDCTextInputControllerBaseLeadingUnderlineLabelTextColor =
-    @"MDCTextInputControllerBaseLeadingUnderlineLabelTextColor";
-static NSString *const MDCTextInputControllerBaseMinimumLinesKey =
-    @"MDCTextInputControllerBaseMinimumLinesKey";
-static NSString *const MDCTextInputControllerBaseNormalColorKey =
-    @"MDCTextInputControllerBaseNormalColorKey";
-static NSString *const MDCTextInputControllerBasePresentationStyleKey =
-    @"MDCTextInputControllerBasePresentationStyleKey";
-static NSString *const MDCTextInputControllerBaseTextInputKey =
-    @"MDCTextInputControllerBaseTextInputKey";
-static NSString *const MDCTextInputControllerBaseTrailingUnderlineLabelFontKey =
-    @"MDCTextInputControllerBaseTrailingUnderlineLabelFontKey";
-static NSString *const MDCTextInputControllerBaseTrailingUnderlineLabelTextColor =
-    @"MDCTextInputControllerBaseTrailingUnderlineLabelTextColor";
-static NSString *const MDCTextInputControllerBaseUnderlineHeightActiveKey =
-    @"MDCTextInputControllerBaseUnderlineHeightActiveKey";
-static NSString *const MDCTextInputControllerBaseUnderlineHeightNormalKey =
-    @"MDCTextInputControllerBaseUnderlineHeightNormalKey";
-static NSString *const MDCTextInputControllerBaseUnderlineViewModeKey =
-    @"MDCTextInputControllerBaseUnderlineViewModeKey";
-
 static inline UIColor *MDCTextInputControllerBaseDefaultInlinePlaceholderTextColorDefault() {
   return [UIColor colorWithWhite:0 alpha:MDCTextInputControllerBaseDefaultHintTextOpacity];
 }
@@ -217,85 +160,7 @@ static UITextFieldViewMode _underlineViewModeDefault = UITextFieldViewModeWhileE
 - (instancetype)initWithCoder:(NSCoder *)aDecoder {
   self = [super init];
   if (self) {
-    _activeColor = [aDecoder decodeObjectOfClass:[UIColor class]
-                                          forKey:MDCTextInputControllerBaseActiveColorKey];
-    _borderFillColor = [aDecoder decodeObjectOfClass:[UIColor class]
-                                              forKey:MDCTextInputControllerBaseBorderFillColorKey];
-    _characterCounter =
-        [aDecoder decodeObjectOfClass:[NSObject<MDCTextInputCharacterCounter> class]
-                               forKey:MDCTextInputControllerBaseCharacterCounterKey];
-    _errorColor = [aDecoder decodeObjectOfClass:[UIColor class]
-                                         forKey:MDCTextInputControllerBaseErrorColorKey];
-    // _floatingPlaceholderActiveColor is not encoded. It was added after we decided to remove this
-    // kind of encoding.
-    _floatingPlaceholderNormalColor =
-        [aDecoder decodeObjectOfClass:[UIColor class]
-                               forKey:MDCTextInputControllerBaseFloatingPlaceholderNormalColorKey];
-    _floatingPlaceholderScale =
-        [aDecoder decodeObjectOfClass:[NSNumber class]
-                               forKey:MDCTextInputControllerBaseFloatingPlaceholderScaleKey];
-    _inlinePlaceholderColor =
-        [aDecoder decodeObjectOfClass:[UIColor class]
-                               forKey:MDCTextInputControllerBaseInlinePlaceholderColorKey];
-    _inlinePlaceholderFont =
-        [aDecoder decodeObjectOfClass:[UIFont class]
-                               forKey:MDCTextInputControllerBaseInlinePlaceholderFontKey];
-    _leadingUnderlineLabelFont =
-        [aDecoder decodeObjectOfClass:[UIFont class]
-                               forKey:MDCTextInputControllerBaseLeadingUnderlineLabelFontKey];
-    _leadingUnderlineLabelTextColor =
-        [aDecoder decodeObjectOfClass:[UIColor class]
-                               forKey:MDCTextInputControllerBaseLeadingUnderlineLabelTextColor];
-    _normalColor = [aDecoder decodeObjectOfClass:[UIColor class]
-                                          forKey:MDCTextInputControllerBaseNormalColorKey];
-    _trailingUnderlineLabelFont =
-        [aDecoder decodeObjectOfClass:[UIFont class]
-                               forKey:MDCTextInputControllerBaseTrailingUnderlineLabelFontKey];
-    _trailingUnderlineLabelTextColor =
-        [aDecoder decodeObjectOfClass:[UIColor class]
-                               forKey:MDCTextInputControllerBaseTrailingUnderlineLabelTextColor];
     [self commonMDCTextInputControllerBaseInitialization];
-    if ([aDecoder containsValueForKey:MDCTextInputControllerBaseUnderlineHeightActiveKey]) {
-      _underlineHeightActive =
-          (CGFloat)[aDecoder decodeDoubleForKey:MDCTextInputControllerBaseUnderlineHeightActiveKey];
-    }
-    if ([aDecoder containsValueForKey:MDCTextInputControllerBaseUnderlineHeightNormalKey]) {
-      _underlineHeightNormal =
-          (CGFloat)[aDecoder decodeDoubleForKey:MDCTextInputControllerBaseUnderlineHeightNormalKey];
-    }
-    if ([aDecoder containsValueForKey:MDCTextInputControllerBaseUnderlineViewModeKey]) {
-      _underlineViewMode = (UITextFieldViewMode)
-        [aDecoder decodeIntegerForKey:MDCTextInputControllerBaseUnderlineViewModeKey];
-    }
-    if ([aDecoder containsValueForKey:MDCTextInputControllerBaseMinimumLinesKey]) {
-      _minimumLines = [aDecoder decodeIntegerForKey:MDCTextInputControllerBaseMinimumLinesKey];
-    }
-    if ([aDecoder containsValueForKey:MDCTextInputControllerBaseDisabledColorKey]) {
-      _disabledColor = [aDecoder decodeObjectOfClass:[UIColor class]
-                                              forKey:MDCTextInputControllerBaseDisabledColorKey];
-    }
-    if ([aDecoder containsValueForKey:MDCTextInputControllerBaseExpandsOnOverflowKey]) {
-      _expandsOnOverflow = [aDecoder decodeBoolForKey:MDCTextInputControllerBaseExpandsOnOverflowKey];
-    }
-    if ([aDecoder containsValueForKey:MDCTextInputControllerBaseTextInputKey]) {
-      _textInput = [aDecoder decodeObjectOfClass:[UIView<MDCTextInput> class]
-                                          forKey:MDCTextInputControllerBaseTextInputKey];
-    }
-    if ([aDecoder containsValueForKey:MDCTextInputControllerBaseFloatingEnabledKey]) {
-      _floatingEnabled = [aDecoder decodeBoolForKey:MDCTextInputControllerBaseFloatingEnabledKey];
-    }
-    if ([aDecoder containsValueForKey:MDCTextInputControllerBaseCharacterCountMaxKey]) {
-      _characterCountMax =
-          [aDecoder decodeIntegerForKey:MDCTextInputControllerBaseCharacterCountMaxKey];
-    }
-    if ([aDecoder containsValueForKey:MDCTextInputControllerBaseCharacterCountViewModeKey]) {
-      _characterCountViewMode =
-          [aDecoder decodeIntegerForKey:MDCTextInputControllerBaseCharacterCountViewModeKey];
-    }
-    if ([aDecoder containsValueForKey:MDCTextInputControllerBaseRoundedCorners]) {
-      _roundedCorners =
-          (UIRectCorner)[aDecoder decodeIntegerForKey:MDCTextInputControllerBaseRoundedCorners];
-    }
 
     // This should happen last because it relies on the state of a ton of properties.
     [self setupInput];
@@ -313,57 +178,6 @@ static UITextFieldViewMode _underlineViewModeDefault = UITextFieldViewModeWhileE
   }
 
   return self;
-}
-
-- (void)encodeWithCoder:(NSCoder *)aCoder {
-  [aCoder encodeObject:self.activeColor forKey:MDCTextInputControllerBaseActiveColorKey];
-  [aCoder encodeObject:self.borderFillColor forKey:MDCTextInputControllerBaseBorderFillColorKey];
-  if ([self.characterCounter conformsToProtocol:@protocol(NSSecureCoding)]) {
-    [aCoder encodeObject:self.characterCounter
-                  forKey:MDCTextInputControllerBaseCharacterCounterKey];
-  }
-  [aCoder encodeInteger:self.characterCountMax
-                 forKey:MDCTextInputControllerBaseCharacterCountMaxKey];
-  [aCoder encodeInteger:self.characterCountViewMode
-                 forKey:MDCTextInputControllerBaseCharacterCountViewModeKey];
-  [aCoder encodeInteger:self.roundedCorners forKey:MDCTextInputControllerBaseRoundedCorners];
-  [aCoder encodeObject:self.disabledColor forKey:MDCTextInputControllerBaseDisabledColorKey];
-  [aCoder encodeObject:self.errorAccessibilityValue
-                forKey:MDCTextInputControllerBaseErrorAccessibilityValueKey];
-  [aCoder encodeObject:self.errorColor forKey:MDCTextInputControllerBaseErrorColorKey];
-  [aCoder encodeObject:self.errorText forKey:MDCTextInputControllerBaseErrorTextKey];
-  [aCoder encodeBool:self.expandsOnOverflow forKey:MDCTextInputControllerBaseExpandsOnOverflowKey];
-  [aCoder encodeBool:self.isFloatingEnabled forKey:MDCTextInputControllerBaseFloatingEnabledKey];
-  // _floatingPlaceholderActiveColor is not encoded. It was added after we decided to remove this
-  // kind of encoding.
-  [aCoder encodeObject:self.floatingPlaceholderNormalColor
-                forKey:MDCTextInputControllerBaseFloatingPlaceholderNormalColorKey];
-  [aCoder encodeObject:self.floatingPlaceholderScale
-                forKey:MDCTextInputControllerBaseFloatingPlaceholderScaleKey];
-  [aCoder encodeObject:self.helperText forKey:MDCTextInputControllerBaseHelperTextKey];
-  [aCoder encodeObject:self.inlinePlaceholderColor
-                forKey:MDCTextInputControllerBaseInlinePlaceholderColorKey];
-  [aCoder encodeObject:self.inlinePlaceholderFont
-                forKey:MDCTextInputControllerBaseInlinePlaceholderFontKey];
-  [aCoder encodeObject:self.leadingUnderlineLabelFont
-                forKey:MDCTextInputControllerBaseLeadingUnderlineLabelFontKey];
-  [aCoder encodeObject:self.leadingUnderlineLabelTextColor
-                forKey:MDCTextInputControllerBaseLeadingUnderlineLabelTextColor];
-  [aCoder encodeInteger:self.minimumLines forKey:MDCTextInputControllerBaseMinimumLinesKey];
-  [aCoder encodeObject:self.normalColor forKey:MDCTextInputControllerBaseNormalColorKey];
-  if ([self.textInput conformsToProtocol:@protocol(NSSecureCoding)]) {
-    [aCoder encodeConditionalObject:self.textInput forKey:MDCTextInputControllerBaseTextInputKey];
-  }
-  [aCoder encodeObject:self.trailingUnderlineLabelFont
-                forKey:MDCTextInputControllerBaseTrailingUnderlineLabelFontKey];
-  [aCoder encodeObject:self.trailingUnderlineLabelTextColor
-                forKey:MDCTextInputControllerBaseTrailingUnderlineLabelTextColor];
-  [aCoder encodeDouble:(double)self.underlineHeightActive
-                forKey:MDCTextInputControllerBaseUnderlineHeightActiveKey];
-  [aCoder encodeDouble:(double)self.underlineHeightNormal
-                forKey:MDCTextInputControllerBaseUnderlineHeightNormalKey];
-  [aCoder encodeInteger:self.underlineViewMode
-                 forKey:MDCTextInputControllerBaseUnderlineViewModeKey];
 }
 
 - (instancetype)copyWithZone:(__unused NSZone *)zone {
@@ -1497,12 +1311,6 @@ static UITextFieldViewMode _underlineViewModeDefault = UITextFieldViewModeWhileE
 
 + (void)setUnderlineViewModeDefault:(UITextFieldViewMode)underlineViewModeDefault {
   _underlineViewModeDefault = underlineViewModeDefault;
-}
-
-#pragma mark - NSSecureCoding
-
-+ (BOOL)supportsSecureCoding {
-  return YES;
 }
 
 #pragma mark - Layout